Bug fix  placing two orders in rapid successfully will lead to one of them being null and the system crashing as we check that condition.
The issue: when calling the syscall to place a new order, we use a global order structure in the venueActivity class. This is used to pass parameters to the syscall and the code uses it when the syscall returns later with a response. By that time however, if another order was placed the structure would have been reset and resulted in placing a null order in the orders list (i've been trying to chase this issue down for a while). To fix his we simply remove the global variable and replace with a local passed as parameter to the syscall code. That code now adds the order to the application order list. The code that adds the order is synchronized, which guarantees no two threads will step on each other tying to add an order.

Testing: try to place orders as fast as possible. Everything should work instead of crashing the app.

git-svn-id: http://192.168.2.90/svn/Bartsy-Andriod/Bartsy@258 c3b44e55-a100-42f6-9c86-32dbf579054b